LayoutTests/media/video-load-readyState.html

<video controls></video>
<script src=video-test.js></script>
<script>
test("video.readyState == HTMLMediaElement.DATA_UNAVAILABLE");
waitForEventAndTest("loadstart", "video.readyState == HTMLMediaElement.DATA_UNAVAILABLE");
waitForEventAndTest("loadedmetadata", "video.readyState == HTMLMediaElement.DATA_UNAVAILABLE");
waitForEventAndTest("loadedfirstframe", "video.readyState == HTMLMediaElement.CAN_SHOW_CURRENT_FRAME");
waitForEventAndTest("canshowcurrentframe", "video.readyState == HTMLMediaElement.CAN_SHOW_CURRENT_FRAME");
waitForEventTestAndEnd("canplaythrough", "video.readyState == HTMLMediaElement.CAN_PLAY_THROUGH");
video.src = "content/test.mp4";
</script>
